requirements.txt
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
bech32==1.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
cosmospy_protobuf==0.0.2
decorator==5.1.1
ecdsa==0.17.0
eventkit==1.0.0
executing==1.2.0
greenlet==2.0.2
grpcio==1.47.2
h11==0.12.0
hdwallets==0.1.2
httpcore==0.15.0
httpx==0.23.0
ib-insync==0.9.81
ibdb==0.0.0a4
idna==3.4
ipdb==0.13.11
ipython==8.10.0
jedi==0.18.2
matplotlib-inline==0.1.6
mnemonic==0.20
mospy-wallet==0.3.4
nest-asyncio==1.5.6
numpy==1.24.2
osmosis_protobuf==0.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
protobuf==4.21.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pysha3==1.0.2
python-dotenv==0.21.1
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
